requirements.txt
ffmpeg-python
minio
python-dotenv